@@ -1,7 +1,7 @@
 import 'dart:io';
 
 class AdMobService {
-  String getAdMobAppId() {
+  String? getAdMobAppId() {
     if (Platform.isAndroid) {
       return 'ca-app-pub-8743953761460890~5808507131';
     } else {
@@ -9,7 +9,7 @@ class AdMobService {
     }
   }
 
-  String getMainPageBannerAdId() {
+  String? getMainPageBannerAdId() {
     if (Platform.isAndroid) {
       return 'ca-app-pub-8743953761460890/8051527095';
     } else {
@@ -17,7 +17,7 @@ class AdMobService {
     }
   }
 
-  String getTestBannerAdId() {
+  String? getTestBannerAdId() {
     if (Platform.isAndroid) {
       return 'ca-app-pub-3940256099942544/6300978111';
     } else {
